* commit '7b5dff134de8b6307913edfe1ab41f7edbea21f0':
  Show correct fighters when they are deduced from other fights
  Sort the fights for /api/arena
  Fix of error in previous commit
  Show tournament name of current fight in Battle HTMAPP-153
  Show number of unscheduled fights per tournament in Schedule
  You can now postpone a fight in Battle
  Show if a fight is finished in arena list in Battle
  Added Ruleset implementation for Mexico 2014 / EMAG (European Martial Arts Gathering)
  Source format
  Renamed ArenaList to Schedule
  timerUpdate and messageUpdate are working again
  Start Fight button is working again in Battle
  Show list of upcoming fights in Battle
  Show tournament name in arena list of Battle
  Removed fighterAParticipant and fighterBParticipant from Fight
  Show event title and fight times&names in Battle arena overview

+++ htm-lib/src/main/scala/nl/malienkolders/htm/lib/rulesets/Ruleset.scala
@@ -25,7 +25,7 @@ abstract class Scores {
   }
 }
 
-case class FightProperties(timeLimit: Long, breakAt: Long, breakDuration: Long, timeBetweenFights: Long, exchangeLimit: Long)
+case class FightProperties(timeLimit: Long, breakAt: Long, breakDuration: Long, timeBetweenFights: Long, exchangeLimit: Int)
 
 trait Ruleset {
